Move fireGlobalEventEvent module
