+++ app/ui-master/client/body.js
@@ -7,12 +7,13 @@ import { Template } from 'meteor/templating';
 
 import { t } from '../../utils/client';
 import { chatMessages } from '../../ui';
-import { Layout, popover, fireGlobalEvent, RoomManager } from '../../ui-utils';
+import { Layout, popover, RoomManager } from '../../ui-utils';
 import { settings } from '../../settings';
 import { ChatSubscription } from '../../models';
 import './body.html';
 import { imperativeModal } from '../../../client/lib/imperativeModal';
 import GenericModal from '../../../client/components/GenericModal';
+import { fireGlobalEvent } from '../../../client/lib/fireGlobalEvent';
 
 Template.body.onRendered(function() {
 	new Clipboard('.clipboard');
