+++ app/ui/client/components/header/header.js
@@ -1,6 +1,7 @@
 import { Template } from 'meteor/templating';
 
-import { fireGlobalEvent } from '../../../../ui-utils';
+import { fireGlobalEvent } from '../../../../../client/lib/fireGlobalEvent';
+
 import './header.html';
 
 Template.header.helpers({
