+++ client/startup/userStatusManuallySet.ts
@@ -1,8 +1,8 @@
 import { Meteor } from 'meteor/meteor';
 
 import { callbacks } from '../../app/callbacks/client';
-import { fireGlobalEvent } from '../../app/ui-utils/client';
 import { UserStatus } from '../../definition/UserStatus';
+import { fireGlobalEvent } from '../lib/fireGlobalEvent';
 
 /* fire user state change globally, to listen on desktop electron client */
 Meteor.startup(() => {
